backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "default-project"
version = "1.0.0"
description = "A short description of your project."
authors = [{ name = "John Scolaro", email = "johnscolaro95@gmail.com" }]
license = { file = "LICENSE" }
readme = "README.md"
keywords = ["strava", "activestatistics"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.13"
dependencies = [
    "boto3==1.34.64",
    "boto3-stubs[essential]==1.34.64",
    "fastapi[standard]==0.115.6",
    "mangum==0.19.0",
    "pandas==2.2.3",
    "pillow==11.0.0",
    "plotly==5.24.1",
    "plotly-calplot==0.1.20",
    "polyline==2.0.2",
    "pyinstrument==5.0.1",
    "sentry-sdk==2.20.0",
    "stravalib==2.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.8.2",
    "mypy==1.9.0",
    "pytest==8.1.1",
    "freezegun==1.4.0",
    "pre-commit==3.6.2",
    "factory-boy==3.3.0",
    "moto==5.0.3",
    "setuptools==75.6.0",
    "types-requests==2.32.0.20241016",
]

[tool.ruff]
lint.select = ["E", "F", "W", "C", "N"]

[tool.mypy]
ignore_missing_imports = true
